backend/shared/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pos-common"
version = "0.1.0"
description = "Shared library for pOS backend services"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "aio-pika>=9.4.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
]

[tool.setuptools.packages.find]
include = ["pos_common*"]
